backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
sqlmodel
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
mutagen
ytmusicapi
alembic
psutil
watchdog
yt-dlp
google-auth
requests
